requirements.txt
markdown-it-py==4.0.0
python-docx==1.2.0
pytest==9.0.3
